feat: [KAN-99] #comment ✅ 0.16.3 업그레이드 ( 원복 ) - 25.02.23 #close #time 1h +review SR @313devops
추신 : 맨앞에 붙일 수 있는 구분
•fix : Bug 류 처리 시 ( patch version up )
•feat : 신규 기능 류 처리 시 ( minor version up )
•perf : 메이저 기능 류 처리 시 ( major version up )

코드리뷰는 코드를 통해 소통하고, 서로의 경험을 통해 배우면서 성장 할 수 있는 기회입니다.
나를 위해, 나의 동료를 위해, 우리 팀을 위해 리뷰 시간을 사용해 주세요.
•너무 많은 코드를 한꺼번에 리뷰하지 마세요.
•최소한의 설명과 함께 중점적으로 봐야 할 곳을 알려 주세요.
•리뷰 후 Comment를 남겨 주세요.
•리뷰는 빠른 시간에 응답해 주세요.
•지적 보다는 대안을 제시해 주세요.

[ Frontend-DOC::Java-Service-Tree-Framework(JSTF)::313DEVGRP ]

[Requirement Manage] http://www.a-rms.net
[Document] http://www.313.co.kr:3010
[Wiki] http://www.313.co.kr/confluence
[IssueTracker] http://www.313.co.kr/jira
[GitLab] http://www.313.co.kr/gitlab
[Redmine] http://www.313.co.kr:3000/

[VersionControl] https://github.com/313devgrp
[CodeReview] http://www.313.co.kr/fecru
[CICD-Deploy] http://www.313.co.kr/jenkins
[BuildManager] http://www.313.co.kr/spinnaker
[ArtifactManager] http://www.313.co.kr/nexus
[CodeAnalysis] http://www.313.co.kr/sonar

[Docker Hub] https://hub.docker.com/u/313devgrp
[Kubernetes] http://www.313.co.kr:31380
[DockerSwarm] http://www.313.co.kr/portainer/#!/auth

[DB Admin] http://www.313.co.kr/phpmyadmin/
[S3 Admin] http://www.313.co.kr/minio/login
[MEM Admin] http://www.313.co.kr/redis/
[NoSql Index] http://www.313.co.kr/elasticsearch/_cat/indices
[NoSql Node] http://www.313.co.kr/elasticsearch/_nodes?pretty=true
[Kibana] http://www.313.co.kr/kibana/app/home#/
[LogStash] http://www.313.co.kr/logstash/_node/stats?pretty
[ZipKin] http://www.313.co.kr/zipkin/
[ElasticHQ] http://www.313.co.kr/elastichq/#!/
[Grafana] http://www.313.co.kr/grafana
[Kafka] http://www.313.co.kr/kafka

[NAS] http://www.313.co.kr:5000/webman/index.cgi
[Mail] http://www.313.co.kr/mail/
[Auth] http://www.313.co.kr/auth/
[RDP] http://www.313.co.kr/guacamole/#/
[DrawIO] http://www.313.co.kr/drawio
[DrawDB] http://www.313.co.kr/drawdb

+++ packages/frontend/core/src/components/page-list/operation-cell.tsx
@@ -1,28 +1,31 @@
 import {
   IconButton,
   Menu,
-  MenuIcon,
   MenuItem,
   toast,
   useConfirmModal,
+  usePromptModal,
 } from '@affine/component';
-import { useAppSettingHelper } from '@affine/core/hooks/affine/use-app-setting-helper';
-import { useBlockSuiteMetaHelper } from '@affine/core/hooks/affine/use-block-suite-meta-helper';
-import { useTrashModalHelper } from '@affine/core/hooks/affine/use-trash-modal-helper';
-import { useCatchEventCallback } from '@affine/core/hooks/use-catch-event-hook';
-import { track } from '@affine/core/mixpanel';
-import { FavoriteService } from '@affine/core/modules/favorite';
-import { CompatibleFavoriteItemsAdapter } from '@affine/core/modules/properties';
+import { useBlockSuiteMetaHelper } from '@affine/core/components/hooks/affine/use-block-suite-meta-helper';
+import { useCatchEventCallback } from '@affine/core/components/hooks/use-catch-event-hook';
+import { WorkspaceDialogService } from '@affine/core/modules/dialogs';
+import { DocsService } from '@affine/core/modules/doc';
+import {
+  CompatibleFavoriteItemsAdapter,
+  FavoriteService,
+} from '@affine/core/modules/favorite';
+import { FeatureFlagService } from '@affine/core/modules/feature-flag';
 import { WorkbenchService } from '@affine/core/modules/workbench';
+import { WorkspaceService } from '@affine/core/modules/workspace';
 import type { Collection, DeleteCollectionInfo } from '@affine/env/filter';
 import { useI18n } from '@affine/i18n';
+import { track } from '@affine/track';
+import type { DocMeta } from '@blocksuite/affine/store';
 import {
   DeleteIcon,
   DeletePermanentlyIcon,
   DuplicateIcon,
   EditIcon,
-  FavoritedIcon,
-  FavoriteIcon,
   FilterIcon,
   FilterMinusIcon,
   InformationIcon,
@@ -32,25 +35,19 @@ import {
   ResetIcon,
   SplitViewIcon,
 } from '@blocksuite/icons/rc';
-import type { DocMeta } from '@blocksuite/store';
-import {
-  useLiveData,
-  useService,
-  useServices,
-  WorkspaceService,
-} from '@toeverything/infra';
+import { useLiveData, useService, useServices } from '@toeverything/infra';
+import type { MouseEvent } from 'react';
 import { useCallback, useState } from 'react';
 
 import type { CollectionService } from '../../modules/collection';
-import { InfoModal } from '../affine/page-properties';
 import { usePageHelper } from '../blocksuite/block-suite-page-list/utils';
+import { IsFavoriteIcon } from '../pure/icons';
 import { FavoriteTag } from './components/favorite-tag';
 import * as styles from './list.css';
 import { DisablePublicSharing, MoveToTrash } from './operation-menu-items';
 import { CreateOrEditTag } from './tags/create-tag';
 import type { TagMeta } from './types';
 import { ColWrapper } from './utils';
-import { useEditCollection, useEditCollectionName } from './view';
 
 const tooltipSideTop = { side: 'top' as const };
 const tooltipSideTopAlignEnd = { side: 'top' as const, align: 'end' as const };
@@ -67,37 +64,65 @@ export const PageOperationCell = ({
   onRemoveFromAllowList,
 }: PageOperationCellProps) => {
   const t = useI18n();
-  const currentWorkspace = useService(WorkspaceService).workspace;
-  const { appSettings } = useAppSettingHelper();
-  const { setTrashModal } = useTrashModalHelper(currentWorkspace.docCollection);
-  const [openDisableShared, setOpenDisableShared] = useState(false);
-  const favAdapter = useService(CompatibleFavoriteItemsAdapter);
+  const {
+    featureFlagService,
+    workspaceService,
+    compatibleFavoriteItemsAdapter: favAdapter,
+    workbenchService,
+  } = useServices({
+    FeatureFlagService,
+    WorkspaceService,
+    CompatibleFavoriteItemsAdapter,
+    WorkbenchService,
+  });
+  const enableSplitView = useLiveData(
+    featureFlagService.flags.enable_multi_view.$
+  );
+  const currentWorkspace = workspaceService.workspace;
   const favourite = useLiveData(favAdapter.isFavorite$(page.id, 'doc'));
-  const workbench = useService(WorkbenchService).workbench;
-  const { duplicate } = useBlockSuiteMetaHelper(currentWorkspace.docCollection);
+  const workbench = workbenchService.workbench;
+  const { duplicate } = useBlockSuiteMetaHelper();
+  const docRecord = useLiveData(useService(DocsService).list.doc$(page.id));
   const blocksuiteDoc = currentWorkspace.docCollection.getDoc(page.id);
 
-  const [openInfoModal, setOpenInfoModal] = useState(false);
+  const workspaceDialogService = useService(WorkspaceDialogService);
   const onOpenInfoModal = useCallback(() => {
-    track.$.docInfoPanel.$.open();
-    setOpenInfoModal(true);
-  }, []);
+    if (blocksuiteDoc?.id) {
+      track.$.docInfoPanel.$.open();
+      workspaceDialogService.open('doc-info', { docId: blocksuiteDoc.id });
+    }
+  }, [blocksuiteDoc?.id, workspaceDialogService]);
 
   const onDisablePublicSharing = useCallback(() => {
+    // TODO(@EYHN): implement disable public sharing
     toast('Successfully disabled', {
       portal: document.body,
     });
   }, []);
 
+  const { openConfirmModal } = useConfirmModal();
+
   const onRemoveToTrash = useCallback(() => {
+    if (!docRecord) {
+      return;
+    }
     track.allDocs.list.docMenu.deleteDoc();
 
-    setTrashModal({
-      open: true,
-      pageIds: [page.id],
-      pageTitles: [page.title],
+    openConfirmModal({
+      title: t['com.affine.moveToTrash.confirmModal.title'](),
+      description: t['com.affine.moveToTrash.confirmModal.description']({
+        title: docRecord.title$.value || t['Untitled'](),
+      }),
+      cancelText: t['com.affine.confirmModal.button.cancel'](),
+      confirmText: t.Delete(),
+      confirmButtonOptions: {
+        variant: 'error',
+      },
+      onConfirm: () => {
+        docRecord.moveToTrash();
+      },
     });
-  }, [page.id, page.title, setTrashModal]);
+  }, [docRecord, openConfirmModal, t]);
 
   const onOpenInSplitView = useCallback(() => {
     track.allDocs.list.docMenu.openInSplitView();
@@ -114,8 +139,8 @@ export const PageOperationCell = ({
     favAdapter.toggle(page.id, 'doc');
     toast(
       status
-        ? t['com.arms.toastMessage.removedFavorites']()
-        : t['com.arms.toastMessage.addedFavorites']()
+        ? t['com.affine.toastMessage.removedFavorites']()
+        : t['com.affine.toastMessage.addedFavorites']()
     );
   }, [page.id, favAdapter, t]);
 
@@ -144,85 +169,39 @@ export const PageOperationCell = ({
       {page.isPublic && (
         <DisablePublicSharing
           data-testid="disable-public-sharing"
-          onSelect={() => {
-            setOpenDisableShared(true);
-          }}
+          onSelect={onDisablePublicSharing}
         />
       )}
       {isInAllowList && (
         <MenuItem
           onClick={handleRemoveFromAllowList}
-          preFix={
-            <MenuIcon>
-              <FilterMinusIcon />
-            </MenuIcon>
-          }
+          prefixIcon={<FilterMinusIcon />}
         >
           {t['Remove special filter']()}
         </MenuItem>
       )}
       <MenuItem
         onClick={onToggleFavoritePageOption}
-        preFix={
-          <MenuIcon>
-            {favourite ? (
-              <FavoritedIcon style={{ color: 'var(--affine-primary-color)' }} />
-            ) : (
-              <FavoriteIcon />
-            )}
-          </MenuIcon>
-        }
+        prefixIcon={<IsFavoriteIcon favorite={favourite} />}
       >
         {favourite
-          ? t['com.arms.favoritePageOperation.remove']()
-          : t['com.arms.favoritePageOperation.add']()}
+          ? t['com.affine.favoritePageOperation.remove']()
+          : t['com.affine.favoritePageOperation.add']()}
       </MenuItem>
-      {runtimeConfig.enableInfoModal ? (
-        <MenuItem
-          onClick={onOpenInfoModal}
-          preFix={
-            <MenuIcon>
-              <InformationIcon />
-            </MenuIcon>
-          }
-        >
-          {t['com.arms.page-properties.page-info.view']()}
+      <MenuItem onClick={onOpenInfoModal} prefixIcon={<InformationIcon />}>
+        {t['com.affine.page-properties.page-info.view']()}
+      </MenuItem>
+      <MenuItem onClick={onOpenInNewTab} prefixIcon={<OpenInNewIcon />}>
+        {t['com.affine.workbench.tab.page-menu-open']()}
+      </MenuItem>
+      {BUILD_CONFIG.isElectron && enableSplitView ? (
+        <MenuItem onClick={onOpenInSplitView} prefixIcon={<SplitViewIcon />}>
+          {t['com.affine.workbench.split-view.page-menu-open']()}
         </MenuItem>
       ) : null}
 
-      <MenuItem
-        onClick={onOpenInNewTab}
-        preFix={
-          <MenuIcon>
-            <OpenInNewIcon />
-          </MenuIcon>
-        }
-      >
-        {t['com.arms.workbench.tab.page-menu-open']()}
-      </MenuItem>
-
-      {environment.isDesktop && appSettings.enableMultiView ? (
-        <MenuItem
-          onClick={onOpenInSplitView}
-          preFix={
-            <MenuIcon>
-              <SplitViewIcon />
-            </MenuIcon>
-          }
-        >
-          {t['com.arms.workbench.split-view.page-menu-open']()}
-        </MenuItem>
-      ) : null}
-
-      <MenuItem
-        preFix={
-          <MenuIcon>
-            <DuplicateIcon />
-          </MenuIcon>
-        }
-        onSelect={onDuplicate}
-      >
-        {t['com.arms.header.option.duplicate']()}
+      <MenuItem prefixIcon={<DuplicateIcon />} onSelect={onDuplicate}>
+        {t['com.affine.header.option.duplicate']()}
       </MenuItem>
 
       <MoveToTrash data-testid="move-to-trash" onSelect={onRemoveToTrash} />
@@ -250,18 +229,6 @@ export const PageOperationCell = ({
           </IconButton>
         </Menu>
       </ColWrapper>
-      {blocksuiteDoc ? (
-        <InfoModal
-          open={openInfoModal}
-          onOpenChange={setOpenInfoModal}
-          docId={blocksuiteDoc.id}
-        />
-      ) : null}
-      <DisablePublicSharing.DisablePublicSharingModal
-        onConfirm={onDisablePublicSharing}
-        open={openDisableShared}
-        onOpenChange={setOpenDisableShared}
-      />
     </>
   );
 };
@@ -282,10 +249,10 @@ export const TrashOperationCell = ({
     e => {
       e.preventDefault();
       openConfirmModal({
-        title: `${t['com.arms.trashOperation.deletePermanently']()}?`,
-        description: t['com.arms.trashOperation.deleteDescription'](),
+        title: `${t['com.affine.trashOperation.deletePermanently']()}?`,
+        description: t['com.affine.trashOperation.deleteDescription'](),
         cancelText: t['Cancel'](),
-        confirmText: t['com.arms.trashOperation.delete'](),
+        confirmText: t['com.affine.trashOperation.delete'](),
         confirmButtonOptions: {
           variant: 'error',
         },
@@ -306,7 +273,7 @@ export const TrashOperationCell = ({
   return (
     <ColWrapper flex={1}>
       <IconButton
-        tooltip={t['com.arms.trashOperation.restoreIt']()}
+        tooltip={t['com.affine.trashOperation.restoreIt']()}
         tooltipOptions={tooltipSideTop}
         data-testid="restore-page-button"
         style={{ marginRight: '12px' }}
@@ -316,7 +283,7 @@ export const TrashOperationCell = ({
         <ResetIcon />
       </IconButton>
       <IconButton
-        tooltip={t['com.arms.trashOperation.deletePermanently']()}
+        tooltip={t['com.affine.trashOperation.deletePermanently']()}
         tooltipOptions={tooltipSideTopAlignEnd}
         data-testid="delete-page-button"
         onClick={onConfirmPermanentlyDelete}
@@ -342,46 +309,63 @@ export const CollectionOperationCell = ({
   info,
 }: CollectionOperationCellProps) => {
   const t = useI18n();
-
-  const favAdapter = useService(CompatibleFavoriteItemsAdapter);
-  const docCollection = useService(WorkspaceService).workspace.docCollection;
+  const {
+    compatibleFavoriteItemsAdapter: favAdapter,
+    workspaceService,
+    workspaceDialogService,
+  } = useServices({
+    CompatibleFavoriteItemsAdapter,
+    WorkspaceService,
+    WorkspaceDialogService,
+  });
+  const docCollection = workspaceService.workspace.docCollection;
   const { createPage } = usePageHelper(docCollection);
   const { openConfirmModal } = useConfirmModal();
   const favourite = useLiveData(
     favAdapter.isFavorite$(collection.id, 'collection')
   );
 
-  const { open: openEditCollectionModal, node: editModal } =
-    useEditCollection();
+  const { openPromptModal } = usePromptModal();
 
-  const { open: openEditCollectionNameModal, node: editNameModal } =
-    useEditCollectionName({
-      title: t['com.arms.editCollection.renameCollection'](),
-    });
+  const handlePropagation = useCallback((event: MouseEvent) => {
+    event.preventDefault();
+    event.stopPropagation();
+  }, []);
 
-  const handleEditName = useCallback(() => {
-    // use openRenameModal if it is in the sidebar collection list
-    openEditCollectionNameModal(collection.name)
-      .then(name => {
-        return service.updateCollection(collection.id, collection => ({
-          ...collection,
-          name,
-        }));
-      })
-      .catch(err => {
-        console.error(err);
+  const handleEditName = useCallback(
+    (event: MouseEvent) => {
+      handlePropagation(event);
+      openPromptModal({
+        title: t['com.affine.editCollection.renameCollection'](),
+        label: t['com.affine.editCollectionName.name'](),
+        inputOptions: {
+          placeholder: t['com.affine.editCollectionName.name.placeholder'](),
+        },
+        confirmText: t['com.affine.editCollection.save'](),
+        cancelText: t['com.affine.editCollection.button.cancel'](),
+        confirmButtonOptions: {
+          variant: 'primary',
+        },
+        onConfirm(name) {
+          service.updateCollection(collection.id, () => ({
+            ...collection,
+            name,
+          }));
+        },
       });
-  }, [collection.id, collection.name, openEditCollectionNameModal, service]);
+    },
+    [collection, handlePropagation, openPromptModal, service, t]
+  );
 
-  const handleEdit = useCallback(() => {
-    openEditCollectionModal(collection)
-      .then(collection => {
-        return service.updateCollection(collection.id, () => collection);
-      })
-      .catch(err => {
-        console.error(err);
+  const handleEdit = useCallback(
+    (event: MouseEvent) => {
+      handlePropagation(event);
+      workspaceDialogService.open('collection-editor', {
+        collectionId: collection.id,
       });
-  }, [openEditCollectionModal, collection, service]);
+    },
+    [handlePropagation, workspaceDialogService, collection.id]
+  );
 
   const handleDelete = useCallback(() => {
     return service.deleteCollection(info, collection.id);
@@ -392,8 +376,8 @@ export const CollectionOperationCell = ({
     favAdapter.toggle(collection.id, 'collection');
     toast(
       status
-        ? t['com.arms.toastMessage.removedFavorites']()
-        : t['com.arms.toastMessage.addedFavorites']()
+        ? t['com.affine.toastMessage.removedFavorites']()
+        : t['com.affine.toastMessage.addedFavorites']()
     );
   }, [favAdapter, collection.id, t]);
 
@@ -404,8 +388,8 @@ export const CollectionOperationCell = ({
 
   const onConfirmAddDocToCollection = useCallback(() => {
     openConfirmModal({
-      title: t['com.arms.collection.add-doc.confirm.title'](),
-      description: t['com.arms.collection.add-doc.confirm.description'](),
+      title: t['com.affine.collection.add-doc.confirm.title'](),
+      description: t['com.affine.collection.add-doc.confirm.description'](),
       cancelText: t['Cancel'](),
       confirmText: t['Confirm'](),
       confirmButtonOptions: {
@@ -417,8 +401,6 @@ export const CollectionOperationCell = ({
 
   return (
     <>
-      {editModal}
-      {editNameModal}
       <ColWrapper
         hideInSmallContainer
         data-testid="page-list-item-favorite"
@@ -429,14 +411,14 @@ export const CollectionOperationCell = ({
       </ColWrapper>
       <IconButton
         onClick={handleEditName}
-        tooltip={t['com.arms.collection.menu.rename']()}
+        tooltip={t['com.affine.collection.menu.rename']()}
         tooltipOptions={tooltipSideTop}
       >
         <EditIcon />
       </IconButton>
       <IconButton
         onClick={handleEdit}
-        tooltip={t['com.arms.collection.menu.edit']()}
+        tooltip={t['com.affine.collection.menu.edit']()}
         tooltipOptions={tooltipSideTop}
       >
         <FilterIcon />
@@ -447,39 +429,21 @@ export const CollectionOperationCell = ({
             <>
               <MenuItem
                 onClick={onToggleFavoriteCollection}
-                preFix={
-                  <MenuIcon>
-                    {favourite ? (
-                      <FavoritedIcon
-                        style={{ color: 'var(--affine-primary-color)' }}
-                      />
-                    ) : (
-                      <FavoriteIcon />
-                    )}
-                  </MenuIcon>
-                }
+                prefixIcon={<IsFavoriteIcon favorite={favourite} />}
               >
                 {favourite
-                  ? t['com.arms.favoritePageOperation.remove']()
-                  : t['com.arms.favoritePageOperation.add']()}
+                  ? t['com.affine.favoritePageOperation.remove']()
+                  : t['com.affine.favoritePageOperation.add']()}
               </MenuItem>
               <MenuItem
                 onClick={onConfirmAddDocToCollection}
-                preFix={
-                  <MenuIcon>
-                    <PlusIcon />
-                  </MenuIcon>
-                }
+                prefixIcon={<PlusIcon />}
               >
                 {t['New Page']()}
               </MenuItem>
               <MenuItem
                 onClick={handleDelete}
-                preFix={
-                  <MenuIcon>
-                    <DeleteIcon />
-                  </MenuIcon>
-                }
+                prefixIcon={<DeleteIcon />}
                 type="danger"
                 data-testid="delete-collection"
               >
@@ -545,7 +509,13 @@ export const TagOperationCell = ({
       <IconButton
         tooltip={t['Rename']()}
         tooltipOptions={tooltipSideTop}
-        onClick={() => setOpen(true)}
+        onClick={useCallback(
+          (e: React.MouseEvent<HTMLButtonElement>) => {
+            e.preventDefault();
+            setOpen(true);
+          },
+          [setOpen]
+        )}
       >
         <EditIcon />
       </IconButton>
@@ -554,11 +524,7 @@ export const TagOperationCell = ({
         <Menu
           items={
             <MenuItem
-              preFix={
-                <MenuIcon>
-                  <DeleteIcon />
-                </MenuIcon>
-              }
+              prefixIcon={<DeleteIcon />}
               type="danger"
               onSelect={handleDelete}
               data-testid="delete-tag"
